In this version, libchess was added to support chess games.
Chess games are implemented int GameChess, a son of Game class
to support the types of chess that the only difference is
the duration of time. Then  a Standard Chess is implementeded
in GameStandard using GameChess.

Note that GameStandard and GameChess are not complete yet,
there are some functions to be implemented.